Refactored authentication API configures
