+++ core/cas-server-core-webflow/src/main/java/org/apereo/cas/web/flow/config/CasCoreWebflowConfiguration.java
@@ -245,16 +245,19 @@ public class CasCoreWebflowConfiguration {
     }
 
     @Bean
+    @ConditionalOnMissingBean(name = "clearWebflowCredentialsAction")
     public Action clearWebflowCredentialsAction() {
         return new ClearWebflowCredentialAction();
     }
 
     @Bean
+    @ConditionalOnMissingBean(name = "checkWebAuthenticationRequestAction")
     public Action checkWebAuthenticationRequestAction() {
         return new CheckWebAuthenticationRequestAction(casProperties.getAuthn().getMfa().getContentType());
     }
 
     @Bean
+    @ConditionalOnMissingBean(name = "redirectToServiceAction")
     public Action redirectToServiceAction() {
         return new RedirectToServiceAction(responseBuilderLocator);
     }
