+++ core/cas-server-core/src/test/java/org/apereo/cas/config/CasMultifactorTestAuthenticationEventExecutionPlanConfiguration.java
@@ -3,7 +3,7 @@ package org.apereo.cas.config;
 import org.apereo.cas.TestOneTimePasswordAuthenticationHandler;
 import org.apereo.cas.authentication.AcceptUsersAuthenticationHandler;
 import org.apereo.cas.authentication.AuthenticationEventExecutionPlan;
-import org.apereo.cas.config.support.authentication.AuthenticationEventExecutionPlanConfigurer;
+import org.apereo.cas.authentication.AuthenticationEventExecutionPlanConfigurer;
 import org.springframework.context.annotation.Configuration;
 
 import java.util.HashMap;
