+++ support/cas-server-support-openid/src/main/java/org/apereo/cas/config/support/authentication/OpenIdAuthenticationEventExecutionPlanConfiguration.java
@@ -1,6 +1,7 @@
 package org.apereo.cas.config.support.authentication;
 
 import org.apereo.cas.authentication.AuthenticationEventExecutionPlan;
+import org.apereo.cas.authentication.AuthenticationEventExecutionPlanConfigurer;
 import org.apereo.cas.authentication.AuthenticationHandler;
 import org.apereo.cas.authentication.principal.DefaultPrincipalFactory;
 import org.apereo.cas.authentication.principal.PrincipalFactory;
