+++ support/cas-server-support-saml/src/main/java/org/apereo/cas/config/authentication/support/SamlAuthenticationEventExecutionPlanConfiguration.java
@@ -2,7 +2,7 @@ package org.apereo.cas.config.authentication.support;
 
 import org.apereo.cas.authentication.AuthenticationEventExecutionPlan;
 import org.apereo.cas.authentication.AuthenticationMetaDataPopulator;
-import org.apereo.cas.config.support.authentication.AuthenticationEventExecutionPlanConfigurer;
+import org.apereo.cas.authentication.AuthenticationEventExecutionPlanConfigurer;
 import org.apereo.cas.configuration.CasConfigurationProperties;
 import org.apereo.cas.support.saml.authentication.SamlAuthenticationMetaDataPopulator;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
