+++ webapp/cas-server-webapp-config/src/main/java/org/apereo/cas/config/support/authentication/AcceptUsersAuthenticationEventExecutionPlanConfiguration.java
@@ -2,6 +2,7 @@ package org.apereo.cas.config.support.authentication;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apereo.cas.authentication.AuthenticationEventExecutionPlan;
+import org.apereo.cas.authentication.AuthenticationEventExecutionPlanConfigurer;
 import org.apereo.cas.authentication.AuthenticationHandler;
 import org.apereo.cas.authentication.principal.PrincipalResolver;
 import org.apereo.cas.configuration.CasConfigurationProperties;
